v6.8.4 1. After the update, the software and panel directory permissions will be set to strict mode (it will be restored automatically after modification) 2. Increase the panel comprehensive anti-riot mechanism 3. Enhance the security of the panel session<br> 4. Enhanced panel entry verification mechanism 5. Optimize the terminal 6. Optimize the panel memory release mechanism Note 1: Version 6.8.2/6.9.11 has security issues, please be sure to upgrade to the latest version Note 2: After this update is successful, it will automatically log out and you need to log in again

+++ class/panel_restore.py
@@ -164,7 +164,8 @@ class panel_restore:
             result = self._restore_backup(self._local_file, site_info, args)
         else:
             return public.ExecShell(False,'Currently only supports restoring local, Google storage and AWS S3 backups')
-        os.remove(self._local_file)
+        if os.path.exists(self._local_file):
+            os.remove(self._local_file)
         if result:
             self._progress_rewrite('Recovery failed: {}'.format(str(site_info['site_path'])))
             return result
